fix(库存) Integer to Long

+++ service-goods/service-goods-core/src/main/java/com/wanmi/sbc/goods/mini/service/goods/WxGoodsService.java
@@ -419,7 +419,7 @@ public class WxGoodsService {
             sku.setSalePrice(price);
             sku.setBarcode("樊登读书--barcode");
             sku.setSkuCode("樊登读书--skuCode");
-            sku.setStockNum(wxLiveAssistantGoodsUpdateParam.getWxStock() == null ? 0 : wxLiveAssistantGoodsUpdateParam.getWxStock().intValue());
+            sku.setStockNum(wxLiveAssistantGoodsUpdateParam.getWxStock() == null ? 0L : wxLiveAssistantGoodsUpdateParam.getWxStock());
             skus.add(sku);
         }
         return skus;
